+++ forge-1-16/CHANGELOG.md
@@ -2,6 +2,6 @@
 
 All notable changes to this project will be documented in this file.
 
-# [UNRELEASED] 
+# [1.16.5-1.0.0.beta1] - 20/12/2023
 
-### Introduced 1.16.5 support.
+### Introduced 1.16.5 support for the mid. Currently only has items. Ore generation, blocks, etc. will be added later.
